+++ Editor/Watchers/ScriptableObjectVariableWatcherEditor.cs
@@ -1,0 +1,10 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace UnderLogic.Variables.Watchers.Editor
+{
+    [CustomEditor(typeof(ScriptableObjectVariableWatcher), true)]
+    public class
+        ScriptableObjectVariableWatcherEditor : RuntimeVariableWatcherEditor<ScriptableObject,
+            ScriptableObjectVariable> { }
+}