+++ Runtime/ScriptableObjectVariable.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Variables
+{
+    [CreateAssetMenu(menuName = "Variables/ScriptableObject Variable")]
+    public class ScriptableObjectVariable : RuntimeVariable<ScriptableObject> { }
+}