+++ Runtime/Watchers/ScriptableObjectVariableWatcher.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Variables.Watchers
+{
+    public class
+        ScriptableObjectVariableWatcher : RuntimeVariableWatcher<ScriptableObject, ScriptableObjectVariable> { }
+}