+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.2] - 2022-12-28
+
+### Added
+
+- `ScriptableObjectVariable` variant
+- `ScriptableObjectWatcher` component
+- Icons for variables
+
 ## [0.6.1] - 2022-12-08
 
 ### Fixed